+++ src/plonemeeting/portal/core/content/subscriber.py
@@ -2,11 +2,12 @@
 
 from eea.facetednavigation.layout.interfaces import IFacetedLayout
 from plone import api
+from plone.api.portal import get_registry_record
 from zope.globalrequest import getRequest
 from zope.interface import alsoProvides
 from zope.i18n import translate
 
-from plonemeeting.portal.core import _
+from plonemeeting.portal.core import _, logger
 from plonemeeting.portal.core.config import APP_FOLDER_ID
 from plonemeeting.portal.core.interfaces import IMeetingsFolder
 from plonemeeting.portal.core.utils import create_faceted_folder, set_constrain_types
@@ -37,6 +38,23 @@ def handle_institution_creation(obj, event):
     request = getRequest()
     if request:  # Request can be `None` during test setup
         request.response.redirect(obj.absolute_url())
+
+
+def handle_institution_modified(institution, event):
+    """
+    Initializes categories_mappings by trying to match global categories with fetched categories from iA.Delib.
+    """
+    if hasattr(institution, "delib_categories") and not institution.categories_mappings:
+        logger.info("Initializing default categories mappings by matching with those fetched from iA.Delib.")
+        global_categories = [cat for cat in get_registry_record(name="plonemeeting.portal.core.global_categories")]
+        local_categories = [cat[0] for cat in institution.delib_categories]
+        categories_mappings = []
+        for local_category in local_categories:
+            if local_category in global_categories:
+                categories_mappings.append({"local_category_id": local_category,
+                                            "global_category_id": local_category})
+        institution.categories_mappings = categories_mappings
+        logger.info("{} fetched iA.Delib categories matched.".format(len(categories_mappings)))
 
 
 def institution_state_changed(obj, event):