+++ CHANGES.rst
@@ -7,10 +7,16 @@ Changelog
 
 - Adapted code to receive the smallest JSON possible by using
   include parameters in the json query.
+- Query categories from iA.Delib to populate vocabularies only when loading the edit form.
+  [odelaere]
+- Added automatic initialization of categories mapping.
+  [odelaere]
 - Fixed applying the demo profile at new Plone Site creation time.
   This was due to BrowserLayers still not initialized, in this case we mark the
   `REQUEST` with registred `BrowserLayers` ourselves.
   [gbastien]
+- Added default value for `Institution.meeting_config_id`
+  [odelaere]
 
 
 1.3.3 (2021-06-28)
@@ -22,7 +28,7 @@ Changelog
   [gbastien]
 - Fix type constraints on Folder content type and faceted folders.
   [odelaere]
-- Merged faceted folders in `Institution , `meetings` and `decisions` were
+- Merged faceted folders in `Institution` , `meetings` and `decisions` were
   merged and only `meetings` folder is kept, new faceted behavior
   is a mix of old behaviors.
   [gbastien]
@@ -59,6 +65,7 @@ Changelog
   `Products.PortalTransforms.safe_html.hasScript` function to accept
   `data:image` base64 value.
   [gbastien]
+
 
 
 1.3.1 (2021-04-29)